+++ wifi/1.0/default/wifi_nan_iface.h
@@ -46,56 +46,56 @@ class WifiNanIface : public IWifiNanIface {
   Return<void> registerEventCallback(
       const sp<IWifiNanIfaceEventCallback>& callback,
       registerEventCallback_cb hidl_status_cb) override;
-  Return<void> enableRequest(uint32_t cmd_id,
+  Return<void> enableRequest(uint16_t cmd_id,
                              const NanEnableRequest& msg,
                              enableRequest_cb hidl_status_cb) override;
-  Return<void> disableRequest(uint32_t cmd_id,
+  Return<void> disableRequest(uint16_t cmd_id,
                               disableRequest_cb hidl_status_cb) override;
-  Return<void> publishRequest(uint32_t cmd_id,
+  Return<void> publishRequest(uint16_t cmd_id,
                               const NanPublishRequest& msg,
                               publishRequest_cb hidl_status_cb) override;
   Return<void> publishCancelRequest(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const NanPublishCancelRequest& msg,
       publishCancelRequest_cb hidl_status_cb) override;
-  Return<void> subscribeRequest(uint32_t cmd_id,
+  Return<void> subscribeRequest(uint16_t cmd_id,
                                 const NanSubscribeRequest& msg,
                                 subscribeRequest_cb hidl_status_cb) override;
   Return<void> subscribeCancelRequest(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const NanSubscribeCancelRequest& msg,
       subscribeCancelRequest_cb hidl_status_cb) override;
   Return<void> transmitFollowupRequest(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const NanTransmitFollowupRequest& msg,
       transmitFollowupRequest_cb hidl_status_cb) override;
-  Return<void> configRequest(uint32_t cmd_id,
+  Return<void> configRequest(uint16_t cmd_id,
                              const NanConfigRequest& msg,
                              configRequest_cb hidl_status_cb) override;
   Return<void> beaconSdfPayloadRequest(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const NanBeaconSdfPayloadRequest& msg,
       beaconSdfPayloadRequest_cb hidl_status_cb) override;
   Return<void> getVersion(getVersion_cb hidl_status_cb) override;
-  Return<void> getCapabilities(uint32_t cmd_id,
+  Return<void> getCapabilities(uint16_t cmd_id,
                                getCapabilities_cb hidl_status_cb) override;
   Return<void> dataInterfaceCreate(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const hidl_string& iface_name,
       dataInterfaceCreate_cb hidl_status_cb) override;
   Return<void> dataInterfaceDelete(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const hidl_string& iface_name,
       dataInterfaceDelete_cb hidl_status_cb) override;
   Return<void> dataRequestInitiator(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const NanDataPathInitiatorRequest& msg,
       dataRequestInitiator_cb hidl_status_cb) override;
   Return<void> dataIndicationResponse(
-      uint32_t cmd_id,
+      uint16_t cmd_id,
       const NanDataPathIndicationResponse& msg,
       dataIndicationResponse_cb hidl_status_cb) override;
-  Return<void> dataEnd(uint32_t cmd_id,
+  Return<void> dataEnd(uint16_t cmd_id,
                        const NanDataPathEndRequest& msg,
                        dataEnd_cb hidl_status_cb) override;
 
@@ -105,34 +105,34 @@ class WifiNanIface : public IWifiNanIface {
   std::pair<WifiStatus, IfaceType> getTypeInternal();
   WifiStatus registerEventCallbackInternal(
       const sp<IWifiNanIfaceEventCallback>& callback);
-  WifiStatus enableRequestInternal(uint32_t cmd_id,
+  WifiStatus enableRequestInternal(uint16_t cmd_id,
                                    const NanEnableRequest& msg);
-  WifiStatus disableRequestInternal(uint32_t cmd_id);
-  WifiStatus publishRequestInternal(uint32_t cmd_id,
+  WifiStatus disableRequestInternal(uint16_t cmd_id);
+  WifiStatus publishRequestInternal(uint16_t cmd_id,
                                     const NanPublishRequest& msg);
-  WifiStatus publishCancelRequestInternal(uint32_t cmd_id,
+  WifiStatus publishCancelRequestInternal(uint16_t cmd_id,
                                           const NanPublishCancelRequest& msg);
-  WifiStatus subscribeRequestInternal(uint32_t cmd_id,
+  WifiStatus subscribeRequestInternal(uint16_t cmd_id,
                                       const NanSubscribeRequest& msg);
   WifiStatus subscribeCancelRequestInternal(
-      uint32_t cmd_id, const NanSubscribeCancelRequest& msg);
+      uint16_t cmd_id, const NanSubscribeCancelRequest& msg);
   WifiStatus transmitFollowupRequestInternal(
-      uint32_t cmd_id, const NanTransmitFollowupRequest& msg);
-  WifiStatus configRequestInternal(uint32_t cmd_id,
+      uint16_t cmd_id, const NanTransmitFollowupRequest& msg);
+  WifiStatus configRequestInternal(uint16_t cmd_id,
                                    const NanConfigRequest& msg);
   WifiStatus beaconSdfPayloadRequestInternal(
-      uint32_t cmd_id, const NanBeaconSdfPayloadRequest& msg);
+      uint16_t cmd_id, const NanBeaconSdfPayloadRequest& msg);
   std::pair<WifiStatus, NanVersion> getVersionInternal();
-  WifiStatus getCapabilitiesInternal(uint32_t cmd_id);
-  WifiStatus dataInterfaceCreateInternal(uint32_t cmd_id,
+  WifiStatus getCapabilitiesInternal(uint16_t cmd_id);
+  WifiStatus dataInterfaceCreateInternal(uint16_t cmd_id,
                                          const std::string& iface_name);
-  WifiStatus dataInterfaceDeleteInternal(uint32_t cmd_id,
+  WifiStatus dataInterfaceDeleteInternal(uint16_t cmd_id,
                                          const std::string& iface_name);
   WifiStatus dataRequestInitiatorInternal(
-      uint32_t cmd_id, const NanDataPathInitiatorRequest& msg);
+      uint16_t cmd_id, const NanDataPathInitiatorRequest& msg);
   WifiStatus dataIndicationResponseInternal(
-      uint32_t cmd_id, const NanDataPathIndicationResponse& msg);
-  WifiStatus dataEndInternal(uint32_t cmd_id, const NanDataPathEndRequest& msg);
+      uint16_t cmd_id, const NanDataPathIndicationResponse& msg);
+  WifiStatus dataEndInternal(uint16_t cmd_id, const NanDataPathEndRequest& msg);
 
   std::string ifname_;
   std::weak_ptr<legacy_hal::WifiLegacyHal> legacy_hal_;
